Refactor and use functions
